+++ src/animus/rest/controllers/library/get_folder_controller.py
@@ -1,0 +1,28 @@
+from typing import Annotated
+
+from fastapi import APIRouter, Depends
+
+from animus.core.library.domain.entities.dtos import FolderDto
+from animus.core.library.interfaces import FoldersRepository
+from animus.core.library.use_cases import GetFolderUseCase
+from animus.core.shared.domain.structures import Id
+from animus.pipes.auth_pipe import AuthPipe
+from animus.pipes.database_pipe import DatabasePipe
+from animus.validation.shared import IdSchema
+
+
+class GetFolderController:
+    @staticmethod
+    def handle(router: APIRouter) -> None:
+        @router.get('/folders/{folder_id}', status_code=200, response_model=FolderDto)
+        def _(
+            folder_id: IdSchema,
+            account_id: Annotated[Id, Depends(AuthPipe.get_account_id_from_request)],
+            folders_repository: Annotated[
+                FoldersRepository,
+                Depends(DatabasePipe.get_folders_repository_from_request),
+            ],
+        ) -> FolderDto:
+            use_case = GetFolderUseCase(folders_repository=folders_repository)
+
+            return use_case.execute(account_id=account_id.value, folder_id=folder_id)