+++ src/animus/routers/library/__init__.py
@@ -1,0 +1,3 @@
+from .library_router import LibraryRouter
+
+__all__ = ['LibraryRouter']